third_party/WebKit/LayoutTests/inspector/tracing/tracing-timeline-load-expected.txt

Tests tracing based Timeline save/load functionality.


Running: testNormal

Running: testJSONObjectFormat

Running: testBroken
error: TracingStartedInPage event not found.

Running: testMalformedJSON
error: TracingStartedInPage event not found.
error: TracingStartedInPage event not found.
error: TracingStartedInPage event not found.
